More calibration info in CDF export
Track the number of stars used to estimate DCM.

More information about the reference calibration file (dates).

Command line is stored in CDF.

Do not process if there are no l1 images in the requested time range.

Renamed some ProgramState variables to not refer to l2 (for case that skymap is used).

Store site info and image elevations and azimuths.